Remove Grok for grids
