+++ src/collective/cover/browser/layout.py
@@ -1,13 +1,11 @@
 # -*- coding: utf-8 -*-
 from AccessControl import getSecurityManager
 from Acquisition import aq_inner
-from collective.cover import _
 from collective.cover.controlpanel import ICoverSettings
 from collective.cover.interfaces import IGridSystem
 from collective.cover.tiles.list import IListTile
 from collective.cover.utils import assign_tile_ids
 from collective.cover.vocabularies import TileStylesVocabulary
-from five import grok
 from plone import api
 from plone.app.uuid.utils import uuidToObject
 from plone.dexterity.events import EditBegunEvent
@@ -22,9 +20,6 @@ from zope.event import notify
 from zope.schema.interfaces import IVocabularyFactory
 
 import json
-
-
-grok.templatedir('templates')
 
 
 # TODO: implement EditCancelledEvent and EditFinishedEvent
@@ -311,104 +306,3 @@ class GroupSelect(BrowserView):
         return self.index()
 
 
-class BaseGrid(object):
-
-    """Base class for grid systems."""
-
-    title = u''
-    ncolumns = 0
-
-    row_class = 'row'
-    column_class = 'column'
-
-    def transform(self, layout):
-        for element in layout:
-            if 'type' in element:
-                if element['type'] == 'row':
-                    element['class'] = self.row_class
-                    if 'css-class' in element:
-                        element['class'] += ' {0}'.format(
-                            element['css-class']
-                        )
-                    if 'children' in element:
-                        self.transform(self.columns_formatter(element['children']))
-                if element['type'] == 'group' and 'children' in element:
-                    self.transform(element['children'])
-
-                if element['type'] == 'tile':
-                    element['class'] = 'tile'
-
-    def columns_formatter(self, columns):
-        raise Exception('Must be implemented in the child')
-
-
-class Bootstrap3(BaseGrid, grok.GlobalUtility):
-
-    """Bootstrap 3 grid system (12 columns)."""
-
-    grok.name('bootstrap3')
-    grok.implements(IGridSystem)
-
-    ncolumns = 12
-    title = _(u'Bootstrap 3')
-
-    def columns_formatter(self, columns):
-        prefix = 'col-md-'
-        for column in columns:
-            width = column.get('column-size', 1)
-            column['class'] = self.column_class + ' ' + (prefix + str(width))
-            if 'css-class' in column:
-                column['class'] += ' {0}'.format(
-                    column['css-class']
-                )
-
-        return columns
-
-
-class Bootstrap2(BaseGrid, grok.GlobalUtility):
-
-    """Bootstrap 2 grid system (12 columns)."""
-
-    grok.name('bootstrap2')
-    grok.implements(IGridSystem)
-
-    ncolumns = 12
-    title = _(u'Bootstrap 2')
-
-    def columns_formatter(self, columns):
-        prefix = 'span'
-        for column in columns:
-            width = column.get('column-size', 1)
-            column['class'] = self.column_class + ' ' + (prefix + str(width))
-            if 'css-class' in column:
-                column['class'] += ' {0}'.format(
-                    column['css-class']
-                )
-
-        return columns
-
-
-class Deco16Grid(BaseGrid, grok.GlobalUtility):
-
-    """Deco grid system (16 columns)."""
-
-    grok.name('deco16_grid')
-    grok.implements(IGridSystem)
-
-    title = _(u'Deco (16 columns)')
-    ncolumns = 16
-    column_class = 'cell'
-
-    def columns_formatter(self, columns):
-        w = 'width-'
-        p = 'position-'
-        offset = 0
-        for column in columns:
-            width = column.get('column-size', 1)
-            column['class'] = self.column_class + ' ' + (w + str(width)) + ' ' + (p + str(offset))
-            if 'css-class' in column:
-                column['class'] += ' {0}'.format(
-                    column['css-class']
-                )
-            offset = offset + width
-        return columns
