+++ src/collective/cover/grids.zcml
@@ -1,0 +1,20 @@
+<configure
+    xmlns="http://namespaces.zope.org/zope"
+    i18n_domain="collective.cover">
+
+  <utility
+      name="bootstrap3"
+      factory=".grids.Bootstrap3"
+      />
+
+  <utility
+      name="bootstrap2"
+      factory=".grids.Bootstrap2"
+      />
+
+  <utility
+      name="deco16_grid"
+      factory=".grids.Deco16Grid"
+      />
+
+</configure>
